refactor: extract shared move-reconciliation helpers into collection/reconcile.rs (#577 item 17)
Second sub-step of splitting collection.rs's 5 bundled domains (#577):
`resolve_case_insensitive_path`, `reconcile_moved_songs`, and
`unreachable_watched_roots` — filesystem-reconciliation logic shared
between the scanner and the realtime watcher — move into a
`collection::reconcile` submodule. `resolve_case_insensitive_path`
stays crate-visible via a re-export so player.rs's existing call site
is unaffected; the other two are called through `reconcile::` from
both collection.rs and watcher.rs.
